segfolting but written algo :D

+++ includes/structures.h
@@ -51,7 +51,7 @@ typedef struct	s_antfarn
 	int		*width_level;
 	int		*loadness;
 	size_t	max_width;
-	size_t	len_of_shortest_path;
+	int	len_of_shortest_path;
 	size_t	max_ant_per_step;
 	int		**steps;
 	_Bool	**busy_vertexes;
